Adds WOM role synchronization bot
Initial implementation of a Discord bot that synchronizes roles
based on Wise Old Man (WOM) group memberships.

The bot automates role assignment and removal, creating roles if
they are missing and removing WOM roles if the user is not assigned
to them in the WOM group. It uses a scheduled task to keep roles
in sync and includes a slash command for manual synchronization.

Adds a .gitignore file to ignore venv
Adds a requirements.txt file for dependencies